fix bug visuel

+++ srcs/game/render/draw_color.c
@@ -14,5 +14,5 @@
 
 unsigned int	create_color(int r, int g, int b)
 {
-	return ((0xFF << 24) | (r << 16) | (g << 8) | b);
+	return (((unsigned int)0xFF << 24) | (r << 16) | (g << 8) | b);
 }
